requirements.txt
Pillow>=10.0.0
PyYAML>=6.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.40.0
accelerate>=0.20.0
